test(converters): add CLI entry-point coverage for all three converter scripts
Extract shared helpers from test_converter_roundtrip.py into
converter_test_utils.py so the setup and verify scripts used by the
shell-driven CLI tests can reuse checkpoint creation and assertion
logic without duplication.

Add _cli_test_setup.py, _cli_test_verify_dcp.py,
_cli_test_verify_meg.py, and _cli_test_verify_lora_merge.py to drive
and verify each converter CLI end-to-end.

Expand test_converters.sh with a second block that invokes all three
converter scripts (five invocations total) via subprocess, covering
parse_args()+main() paths that were previously untested: config.yaml
reading, tokenizer fallback branch, --hf-model-name override,
--adapter-only dispatch, and hf_overrides extraction.
